+++ InventoryTools/Logic/Settings/SwitchCraftListsAutomaticallySetting.cs
@@ -6,7 +6,7 @@ namespace InventoryTools.Logic.Settings
 {
     public class SwitchCraftListsAutomaticallySetting : BooleanSetting
     {
-        public override bool DefaultValue { get; set; } = false;
+        public override bool DefaultValue { get; set; } = true;
         public override bool CurrentValue(InventoryToolsConfiguration configuration)
         {
             return configuration.SwitchCraftListsAutomatically;
@@ -22,7 +22,7 @@ namespace InventoryTools.Logic.Settings
 
         public override string HelpText { get; set; } =
             "Should the active craft list automatically change when moving between each craft list? The active craft list will only change if there is an active craft list already selected.";
-        
+
         public override string Version => "1.7.0.0";
 
         public SwitchCraftListsAutomaticallySetting(ILogger<SwitchCraftListsAutomaticallySetting> logger, ImGuiService imGuiService) : base(logger, imGuiService)